pyproject.toml
[project]
name = "xbook"
version = "1.0.0"
description = "A command line tool for the checking and bookings of fitness time slots at X, TU Delft's sports & culture centre."
authors = [{ name = "Fiske Schijlen" }, { name = "Gautham Venkataraman" }]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
  "dateparser ~= 1.2.0",
  "google-api-python-client ~= 2.133.0",
  "requests ~= 2.32.3",
  "rich ~= 13.7.1",
  "typer ~= 0.12.3",
]

[project.scripts]
xbook = "xbook:app"

[tool.uv]
dev-dependencies = [
  "mypy ~= 1.10.0",
  "types-dateparser ~= 1.2.0.20240420",
  "types-requests ~= 2.32.0.20240602",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
